FM1D2D-2858: Export FM model with original file structure (#607)
* FM1D2D-2858: Some model property (definition) refactoring

* FM1D2D-2858: Refactored towards WaterFlowFMModel.FileStructure

* FM1D2D-2858: Refactored towards WaterFlowFMModel.DimrModel

* FM1D2D-2858: Refactored towards WaterFlowFMModel.IO.Export

* FM1D2D-2858: Refactored towards WaterFlowFMModel.IO.ImportOpen

* FM1D2D-2858: Refactored towards WaterFlowFMModel.Eventing

* FM1D2D-2858: Refactored towards WaterFlowFMModel.NHibernate

* FM1D2D-2858: Refactored towards WaterFlowFMModel.Output.Connecting

* FM1D2D-2858: Refactored towards WaterFlowFMModel.Output.Data

* FM1D2D-2858: Refactored towards WaterFlowFMModel.Output.Saving

* FM1D2D-2858: Refactored towards WaterFlowFMModel.Running

* FM1D2D-2858: Refactored towards WaterFlowFMModel.TDMB

* FM1D2D-2858: Moved additional members

* FM1D2D-2858: Retrieve model directory from MDU path or MDU file references

* FM1D2D-2858: Preserve relative MDU dir in destination path

* FM1D2D-2858: Some refactoring of file related properties

* FM1D2D-2858: Allow relative path up as feature group name

* FM1D2D-2858: Log invalid structures file version

* FM1D2D-2858: Write MDU referenced files with original relative path

* FM1D2D-2858: Export FM model to a folder with original file structure

* FM1D2D-2858: Ranorex - fixed export MDU to directory

* FM1D2D-2858: Ranorex - fixed export DIMR config to directory

* FM1D2D-2858: Ranorex - fixed export grid file

* FM1D2D-2866: Review comments

* FM1D2D-2866: Review comments

* FM1D2D-2866: Review comments

Former-commit-id: ab58395e8ede8c2da7a25c91cf7ba9e05beb7e77

+++ src/DeltaShell.Plugins.FMSuite.FlowFM/IO/BndExtForceFile.cs
@@ -200,11 +200,8 @@ namespace DeltaShell.Plugins.FMSuite.FlowFM.IO
             {
                 WriteMeteoExtForceFile(meteoExtForceFileItems);
             }
-            if (allItems.Any() || meteoExtForceFileItems.Any())
-            {
-                modelProperty.SetValueFromString(Path.GetFileName(FilePath));
-            }
-            else
+
+            if (!allItems.Any() && !meteoExtForceFileItems.Any())
             {
                 FileUtils.DeleteIfExists(FilePath);
                 modelProperty.SetValueFromString(string.Empty);
